+++ src/features/authentication/services/apiSignup.ts
@@ -11,7 +11,7 @@ export interface ResData {
   message: string;
 }
 
-export const signup = async (data: Signup) => {
+export const apiSignup = async (data: Signup) => {
   try {
     const res = await axios.post<ResData>(`${API_URL}/api/v1/auth/signup`, data);
