+++ src/features/authentication/utils/isAuthenticated.ts
@@ -1,4 +1,4 @@
-import { getItem } from '../../../utils/localStorage';
+import { getItem } from '../../utils/localStorage';
 
 export const isAuthenticated = (): boolean => {
   const token = getItem('token');
