+++ src/features/movies/BuyOrRent.tsx
@@ -1,5 +1,5 @@
-import CloseModalBtn from '../../ui/CloseModalBtn';
-import Modal from '../../ui/Modal';
+import CloseModalBtn from '../ui/CloseModalBtn';
+import Modal from '../ui/Modal';
 
 interface BuyOrRentProps {
   onClose(): void;
