+++ src/features/movies/RecommendedMovieList.tsx
@@ -5,9 +5,9 @@ import { IoChevronForward } from 'react-icons/io5';
 import { IoChevronBack } from 'react-icons/io5';
 
 import RecommendedMovieItem from './RecommendedMovieItem';
-import Spinner from '../../ui/Spinner';
 
 import { getAllReleases } from './services/apiReleases';
+import Spinner from '../ui/Spinner';
 
 const RecommendedMovieList = (): JSX.Element => {
   const [translateCount, setTranslateCount] = useState(0);
