+++ src/features/ui/CloseModalBtn.tsx
