+++ src/features/ui/Hamburger.tsx
@@ -2,7 +2,8 @@ import { useEffect } from 'react';
 
 import HamburgerMenuList from './HamburgerMenuList';
 import { IoChevronForward } from 'react-icons/io5';
-import { getEmail, isAuthenticated, logout } from '../features/authentication/utils';
+
+import { getEmail, isAuthenticated, logout } from '../authentication/utils';
 
 interface HamburgerMenuProps {
   onShowSignIn(): void;
